feat: show configured public key
Display the public key derived from the locally configured private key on the main screen.

Reuse the existing Base58 and Base64 local private key decoding path for both the wallet keypair and displayed public key.
